requirements.txt
# Core evaluator + Vercel Python /api functions
pandas>=2.1.0
openai>=1.30.0
pydantic>=2.6.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
